+++ idl/todo.thrift
@@ -10,7 +10,7 @@ struct Todo {
     5: i64 userId;
     6: string createdAt;
     7: string diedAt;
-    8: i32 priority;
+    8: i64 priority;
 }
 
 //增加Todo请求
